requirements.txt
requests==2.31.0
pandas>=1.5.0
python-dotenv==1.0.0
openai==1.3.7
ta==0.10.2
python-binance==1.0.19
urllib3<2.0.0  # Use version compatible with LibreSSL
httpx==0.24.1  # Specific version that works with OpenAI client
alpaca-py>=0.8.0
numpy>=1.21.0
python-dateutil>=2.8.2
pytz>=2022.1 